+++ figma.d.ts
@@ -1,4 +1,4 @@
-// Figma Plugin API version 1, update 4
+// Figma Plugin API version 1, update 5
 
 // Global variable with Figma's plugin API.
 declare const figma: PluginAPI;
@@ -23,9 +23,18 @@ interface PluginAPI {
   readonly root: DocumentNode;
   currentPage: PageNode;
 
-  on(type: 'selectionchange' | 'currentpagechange', callback: () => void); // PROPOSED API ONLY
-  once(type: 'selectionchange' | 'currentpagechange', callback: () => void); // PROPOSED API ONLY
-  off(type: 'selectionchange' | 'currentpagechange', callback: () => void); // PROPOSED API ONLY
+  on(
+    type: 'selectionchange' | 'currentpagechange' | 'close',
+    callback: () => void
+  );
+  once(
+    type: 'selectionchange' | 'currentpagechange' | 'close',
+    callback: () => void
+  );
+  off(
+    type: 'selectionchange' | 'currentpagechange' | 'close',
+    callback: () => void
+  );
 
   readonly mixed: symbol;
 
@@ -119,6 +128,7 @@ interface ShowUIOptions {
   visible?: boolean;
   width?: number;
   height?: number;
+  position?: 'default' | 'last' | 'auto'; // PROPOSED API ONLY
 }
 
 interface UIPostMessageOptions {
@@ -141,12 +151,10 @@ interface UIAPI {
   close(): void;
 
   postMessage(pluginMessage: any, options?: UIPostMessageOptions): void;
-  onmessage:
-    | ((pluginMessage: any, props: OnMessageProperties) => void)
-    | undefined;
-  on(type: 'message', callback: MessageEventHandler); // PROPOSED API ONLY
-  once(type: 'message', callback: MessageEventHandler); // PROPOSED API ONLY
-  off(type: 'message', callback: MessageEventHandler); // PROPOSED API ONLY
+  onmessage: MessageEventHandler | undefined;
+  on(type: 'message', callback: MessageEventHandler);
+  once(type: 'message', callback: MessageEventHandler);
+  off(type: 'message', callback: MessageEventHandler);
 }
 
 interface ViewportAPI {
