+++ src/assets/css/ui.css
@@ -197,7 +197,7 @@ body {
   grid-template-columns: 1fr 32px;
   grid-gap: 10px;
   margin: 0;
-  padding: 15px;
+  padding: 6px;
   border-top: 1px solid #e9e9e9;
 }
 
@@ -213,6 +213,10 @@ body {
   width: 4px;
 }
 
+::-webkit-scrollbar:horizontal {
+  height: 4px;
+}
+
 ::-webkit-scrollbar-track {
   background-color: #e9e9e9;
 }
@@ -221,58 +225,6 @@ body {
   background-color: #000000;
 }
 
-/*
-  SETTINGS
-*/
-.settings {
-  padding: 20px;
-  width: 100%;
-  height: 100%;
-  display: grid;
-  grid-template-rows: 1fr auto;
-}
-.settings h4 {
-  margin: 0 0 15px;
-}
-.settings p {
-  color: #999;
-  margin: 0;
-  font-size: 10px;
-}
-.settings .fields {
-  margin-bottom: 20px;
-}
-.settings .footer {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-gap: 20px;
-}
-.settings .footer .remove-all-messages {
-  grid-column: 1 / span 2;
-}
-.settings .colors {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-  grid-gap: 15px;
-  margin-bottom: 30px;
-}
-.settings .colors .color {
-  position: relative;
-  width: 25px;
-  height: 25px;
-  border-radius: 100%;
-  cursor: pointer;
-}
-.settings .colors .color.active::after {
-  content: '';
-  left: -6px;
-  top: -6px;
-  position: absolute;
-  border-radius: 100%;
-  width: 35px;
-  height: 35px;
-  border: 1px solid #000;
-}
 
 .connection {
   display: grid;
