+++ tsconfig.json
@@ -3,6 +3,7 @@
     "target": "es6",
     "jsx": "react",
     "moduleResolution": "node",
-    "allowSyntheticDefaultImports": true
+    "allowSyntheticDefaultImports": true,
+    "lib": ["dom"]
   }
 }
